RSMB is able to wrap and unwrap encapsulated packet. Multiple wireless nodes per forwarder device are supported. Each wireless node connecting via one forwarder share same socket, so I expect problems when they disconnect. It has to be investigated.

+++ rsmb/src/Clients.h
@@ -279,7 +279,7 @@ typedef struct
 	PendingRegistration* pendingRegistration;
 	uint8_t* wirelessNodeId;         /**< Wireless Node ID used in Encapsulation forwarder packet.
 	                                  *< If not NULL client packets will be encapsulated */
-	size_t wirelessNodeIdLen;        /**< Length of Wireless Node ID  */
+	uint8_t wirelessNodeIdLen;       /**< Length of Wireless Node ID  */
 #if !defined(NO_BRIDGE)
 	PendingSubscription* pendingSubscription;
 #endif
